refactor(linux-client): move all code into `firezone-client-tunnel`
This allows is to be shared with the tunnel binary that's packaged
as part of the GUI Client.
