+++ rust/client-tunnel/Cargo.toml
@@ -6,6 +6,21 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = { version = "1.0" }
+# Will be needed to safely revert `/etc/resolv.conf`
+atomicwrites = "0.4.3"
+secrecy = { workspace = true }
+connlib-client-shared = { workspace = true }
+connlib-shared = { workspace = true }
+firezone-cli-utils = { workspace = true }
+tracing = { workspace = true }
+clap = { version = "4.5", features = ["derive",  "env"] }
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
+humantime = "2.1"
+resolv-conf = "0.7.0"
+thiserror = "1.0.57"
+tokio = { version = "1.36", default-features = false, features = ["rt", "macros", "signal"] }
+url = { version = "2.3.1", default-features = false }
 
 [lints]
 workspace = true
