+++ rust/client-tunnel/src/main.rs
@@ -1,3 +1,4 @@
-fn main() {
-    firezone_client_tunnel::run()
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    firezone_client_tunnel::run().await
 }
